+++ plugins/ucc_chat/lib/ucc_chat/models/schema/notification_setting.ex
@@ -2,6 +2,7 @@ defmodule UccChat.Schema.NotificationSetting do
   use UccChat.Shared, :schema
 
   embedded_schema do
+    field :audio_mode, :string, default: "default"
     field :audio, :string, default: "system_default"
     field :desktop, :string, default: "mentions"
     field :duration, :integer, default: nil
@@ -11,7 +12,7 @@ defmodule UccChat.Schema.NotificationSetting do
   end
 
   @fields [
-    :audio, :desktop, :duration, :mobile, :email, :unread_alert
+    :audio, :desktop, :duration, :mobile, :email, :unread_alert, :audio_mode
   ]
 
   def changeset(struct, params \\ %{}) do