+++ plugins/ucc_chat/lib/ucc_chat_web/admin/chat_general.ex
@@ -1,9 +1,15 @@
 defmodule UccChatWeb.Admin.Page.ChatGeneral do
+  @doc """
+  FlexTab panel implementation for ChatGeneral Administration page.
+  """
   use UccAdmin.Page
 
   alias UcxUcc.{Repo, Hooks}
   alias UccChat.Settings.ChatGeneral
 
+  @doc """
+  Callback to add the ChatGeneral page into the administration pages.
+  """
   def add_page do
     new(
       "admin_chat_general",
@@ -16,11 +22,36 @@ defmodule UccChatWeb.Admin.Page.ChatGeneral do
     )
   end
 
+  @doc """
+  Callback to provide the ChatGeneral page rendering bindings.
+  """
   def args(page, user, _sender, socket) do
     {[
       user: Repo.preload(user, Hooks.user_preload([])),
       changeset: ChatGeneral.get |> ChatGeneral.changeset,
     ], user, page, socket}
   end
+
+  @doc """
+  Helper functions to provide HTML select control options for rendering
+  the page.
+  """
+  def options(:notifications), do: [
+    {~g(All messages), "all"},
+    {~g(Mentions), "mentions"},
+    {~g(Nothing), "none"}
+  ]
+
+  def options(:unread_count), do: [
+    {~g(All messages), "all"},
+    {~g(User mentions only), "user"},
+    {~g(Group mentions only), "group"},
+    {~g(User and group mentions only), "user_and_group"}
+  ]
+
+  def options(:unread_count_dm), do: [
+    {~g(All messages), "all"},
+    {~g(Mentions only), "mentions_only"}
+  ]
 
 end